+++ src/main/java/dev/prithvis/blogbackend/service/impl/BlogUserServiceImpl.java
@@ -4,9 +4,9 @@ import dev.prithvis.blogbackend.exceptions.EntityNotFoundException;
 import dev.prithvis.blogbackend.models.BlogUser;
 import dev.prithvis.blogbackend.payloads.BlogUserDTO;
 import dev.prithvis.blogbackend.repository.BlogUserRepository;
-import dev.prithvis.blogbackend.service.repo.BlogUserService;
-import dev.prithvis.blogbackend.utils.BlogUserToDto;
-import dev.prithvis.blogbackend.utils.DTOToBlogUser;
+import dev.prithvis.blogbackend.service.base.BlogUserService;
+import dev.prithvis.blogbackend.utils.converter.BlogUserToDto;
+import dev.prithvis.blogbackend.utils.converter.DTOToBlogUser;
 import dev.prithvis.blogbackend.utils.MessageUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
