+++ src/main/java/dev/prithvis/blogbackend/utils/converter/CategoryToDto.java
@@ -1,4 +1,4 @@
-package dev.prithvis.blogbackend.utils;
+package dev.prithvis.blogbackend.utils.converter;
 
 import dev.prithvis.blogbackend.models.Category;
 import dev.prithvis.blogbackend.payloads.CategoryDTO;
