+++ src/main/java/dev/prithvis/blogbackend/utils/converter/DTOToBlogUser.java
@@ -1,9 +1,9 @@
-package dev.prithvis.blogbackend.utils;
+package dev.prithvis.blogbackend.utils.converter;
 
 import dev.prithvis.blogbackend.models.BlogUser;
 import dev.prithvis.blogbackend.payloads.BlogUserDTO;
-import lombok.NonNull;
 import org.springframework.core.convert.converter.Converter;
+import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Component;
 
 @Component
