+++ src/test/java/dev/prithvis/blogbackend/controller/BlogUsersControllerTest.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectWriter;
 import dev.prithvis.blogbackend.models.BlogUser;
 import dev.prithvis.blogbackend.payloads.BlogUserDTO;
-import dev.prithvis.blogbackend.service.repo.BlogUserService;
-import dev.prithvis.blogbackend.utils.BlogUserToDto;
+import dev.prithvis.blogbackend.service.base.BlogUserService;
+import dev.prithvis.blogbackend.utils.converter.BlogUserToDto;
 import org.json.JSONObject;
 import org.junit.jupiter.api.Test;
 import org.junit.runner.RunWith;
